Backend/student_portal/requirements.txt
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.15.1
bcrypt==3.2.0
djangorestframework-simplejwt) (4.13.2)
python-magic-0.4.27
openpyxl==3.1.2
Pillow==10.0.0
django-filter==23.2.0
django-extensions==3.2.1
django-debug-toolbar==4.0.1